src/ingestion/semantic_scholar/requirements.txt
google-cloud-bigquery
requests
pandas
